Use short month names (e.g. Aug) consistently across all dates

+++ portal-next/src/app/my-shop/page.tsx
@@ -9,7 +9,7 @@ function formatMemberSince(dateString: string | null) {
   if (!dateString) return '—';
   const date = new Date(dateString);
   if (isNaN(date.getTime())) return '—';
-  return date.toLocaleDateString('en-US', { month: 'long', day: 'numeric', year: 'numeric' });
+  return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' });
 }
 
 function formatJoinedDate(dateString: string | null) {
